+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [2.8.8]
 
+### Removed
+
+- Google Test framework experimental discovery.
+
 ## [2.8.7] - 2020-03-03
 
 Improved logging and some extra error handling.